upgrade next.js versions - Make components folder to clean up the landing page file a bit - Updates to eslint and prettier to fix issues with zed editor - Upgrade packages to make dependencies work on latest next.js version
