+++ package.json
@@ -10,37 +10,45 @@
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
-    "@headlessui/react": "^2.0.4",
+    "@headlessui/react": "^2.2.0",
     "@mapbox/rehype-prism": "^0.9.0",
-    "@mdx-js/loader": "^3.0.1",
-    "@mdx-js/react": "^3.0.1",
-    "@next/mdx": "^14.2.3",
-    "@tailwindcss/typography": "^0.5.13",
+    "@mdx-js/loader": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "@next/mdx": "15.1.2",
+    "@tailwindcss/typography": "^0.5.15",
     "@types/mdx": "^2.0.13",
-    "@types/node": "^20.14.2",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@types/node": "^20.17.10",
+    "@types/react": "19.0.2",
+    "@types/react-dom": "19.0.2",
     "@types/webpack-env": "^1.18.5",
-    "autoprefixer": "^10.4.19",
+    "@typescript-eslint/eslint-plugin": "^8.18.2",
+    "@typescript-eslint/parser": "^8.18.2",
+    "autoprefixer": "^10.4.20",
     "cheerio": "1.0.0-rc.12",
     "clsx": "^2.1.1",
     "fast-glob": "^3.3.2",
     "feed": "^4.2.2",
-    "next": "^14.2.3",
+    "next": "15.1.2",
     "next-themes": "^0.2.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "remark-gfm": "^4.0.0",
-    "resend": "^3.3.0",
-    "tailwindcss": "^3.4.4",
-    "typescript": "^5.4.5",
-    "zod": "^3.23.8"
+    "resend": "^4.0.1",
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.7.2",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.3",
-    "prettier": "^3.3.1",
+    "eslint": "^9.17.0",
+    "eslint-config-next": "15.1.2",
+    "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.5.14",
     "sharp": "0.33.1"
+  },
+  "pnpm": {
+    "overrides": {
+      "@types/react": "19.0.2",
+      "@types/react-dom": "19.0.2"
+    }
   }
 }
